requirements.txt
Flask
python-dotenv
scikit-learn
pandas
numpy
requests
gunicorn
pytest